move AdminController to UserController
