+++ Hairdresser/Program.cs
@@ -41,7 +41,6 @@ builder.Services.AddIdentity<ApplicationUser, IdentityRole>(options =>
 
 builder.Services.AddScoped<JWT_Service>();
 builder.Services.AddScoped<IUserService, UserService>();
-builder.Services.AddScoped<IAdminService, AdminService>();
 builder.Services.AddScoped<ITreatmentService, TreatmentService>();
 
 //JWT Authentication
